+++ app/assets/javascripts/interfaces/demos.js
@@ -1,0 +1,6 @@
+//= require prism
+//= require lodash
+//= require interfaces/demo-scripts
+
+// initialize Prism for Turbolinks
+$(document).on('ready page:load', function () { Prism.highlightAll(); });