+++ app/assets/javascripts/interfaces/dependencies.js
@@ -1,0 +1,2 @@
+//= require device
+//= require interfaces/helpdesk-embed