+++ app/assets/javascripts/interfaces/interfaces.js
@@ -1,0 +1,10 @@
+//= require interfaces/interfaces.components
+//= require interfaces/interfaces.modals
+//= require interfaces/interfaces.tabs
+//= require interfaces/interfaces.dropdowns
+//= require interfaces/interfaces.data_api
+//= require interfaces/helpers
+//= require interfaces/utils
+//= require interfaces/interfacesURLForEnv
+//= require interfaces/dependencies
+//= require interfaces/shims