+++ app/assets/javascripts/interfaces/shims.js
@@ -1,0 +1,6 @@
+//= require modernizr
+//= require es5-shim/es5-shim
+//= require html5shiv
+//= require html5shiv-printshiv
+//= require placeholder
+//= require interfaces/plugins