+++ lib/interfaces.rb
@@ -3,3 +3,5 @@ require 'high_voltage'
 require 'pco/url'
 require 'interfaces/engine'
 require 'interfaces/high_voltage'
+require "interfaces/null_person"
+require "interfaces/null_organization"