+++ lib/interfaces/null_organization.rb
@@ -1,0 +1,5 @@
+module Interfaces
+  class NullOrganization
+    def name; end
+  end
+end