+++ test/dummy/app/assets/javascripts/application.js
@@ -12,6 +12,8 @@
 //
 //= require jquery
 //= require jquery_ujs
+//= require react
+//= require react_ujs
 //= require turbolinks
 //= require interfaces/interfaces
 //= require interfaces/demos